refactor: Move `makeHtmlErrorPage` from domain to view
