--- src/app/domain/errors/models/ErrorPageGenerator.ts
@@ -1,12 +1,0 @@
-export interface ErrorPageGeneratorArguments {
-  backgroundColor?: string
-  error?: {
-    message: string
-    details: string
-  }
-}
-
-export type ErrorPageGenerator = ({
-  backgroundColor,
-  error
-}: ErrorPageGeneratorArguments) => string
